First actual async function test passed, with some arm twisting.

+++ kwh-modbus-tests/WindowsFunctions.h
@@ -13,7 +13,7 @@ typedef void *handle;
 class WindowsFunctions
 {
 private:
-	unsigned long long _hProv = NULL;
+	unsigned long _hProv = NULL;
 
 public:
 	WindowsFunctions();
